SVM: Novelty detection using 1-class SVM
Jira: MADLIB-990

This commit includes the new SQL function for one-class SVM, along with
the corresponding python function that is invoked. This commit also includes
the user doc and design doc updates.
In this implementation of a one-class SVM in MADlib, we are piggy-backing
on the existing SVM classification implementation. The input table to a
one-class SVM does not require a dependent variable. We thus create a temporary
copy of the input table, with a new column for the dependent variable, which
is assigned a default value of 1 (a class label). We then call the existing
SVM classification method, with the new temp table as the source table.